Move code to module
