--- lib/ancient_movie.rb
@@ -1,5 +1,0 @@
-class AncientMovie < Movie
-  def to_s
-    "#{title} - old movie (#{year} year)"
-  end
-end
