--- lib/classic_movie.rb
@@ -1,6 +1,0 @@
-class ClassicMovie < Movie
-  def to_s
-    movies_count = movie_collection.filter(director: director).count
-    "#{title} - classic movie, director #{director} (his #{movies_count} another films)"
-  end
-end
