--- lib/modern_movie.rb
@@ -1,5 +1,0 @@
-class ModernMovie < Movie
-  def to_s
-    "#{title} - modern movie: stars #{star_actors.join(', ')}"
-  end
-end
