+++ lib/movie_industry/ancient_movie.rb
@@ -1,0 +1,7 @@
+module MovieIndustry
+  class AncientMovie < MovieIndustry::Movie
+    def to_s
+      "#{title} - old movie (#{year} year)"
+    end
+  end
+end
