+++ lib/movie_industry/cashbox.rb
@@ -1,0 +1,23 @@
+module MovieIndustry
+  module Cashbox
+    require 'money'
+
+    attr_reader :cash
+
+    def setup_cashbox
+      Money.locale_backend = :currency
+      Money.rounding_mode = BigDecimal::ROUND_HALF_EVEN
+      @cash = Money.new(0, 'USD')
+    end
+
+    def enroll(amount)
+      @cash += amount
+    end
+
+    def take(who = nil)
+      raise 'This is a Robbery!' unless who == 'Bank'
+
+      @cash = Money.new(0)
+    end
+  end
+end
