+++ lib/movie_industry/classic_movie.rb
@@ -1,0 +1,8 @@
+module MovieIndustry
+  class ClassicMovie < MovieIndustry::Movie
+    def to_s
+      movies_count = movie_collection.filter(director: director).count
+      "#{title} - classic movie, director #{director} (his #{movies_count} another films)"
+    end
+  end
+end
