+++ lib/movie_industry/modern_movie.rb
@@ -1,0 +1,7 @@
+module MovieIndustry
+  class ModernMovie < MovieIndustry::Movie
+    def to_s
+      "#{title} - modern movie: stars #{star_actors.join(', ')}"
+    end
+  end
+end
