+++ lib/movie_industry/new_movie.rb
@@ -1,0 +1,7 @@
+module MovieIndustry
+  class NewMovie < MovieIndustry::Movie
+    def to_s
+      "#{title} - new movie, released #{Date.today.year - year} years ago!"
+    end
+  end
+end
