--- lib/new_movie.rb
@@ -1,5 +1,0 @@
-class NewMovie < Movie
-  def to_s
-    "#{title} - new movie, released #{Date.today.year - year} years ago!"
-  end
-end
