+++ spec/movie_industry/new_movie_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe NewMovie do
+RSpec.describe MovieIndustry::NewMovie do
   describe '#to_s' do
     subject { movie.to_s }
