+++ spec/spec_helper.rb
@@ -2,7 +2,7 @@ require 'rubygems'
 require 'bundler/setup'
 Bundler.require(:default, :test, :development)
 require 'csv'
-require './lib/movie.rb'
+require './lib/movie_industry/movie.rb'
 Dir[File.expand_path(File.join(File.dirname(__FILE__), '..', 'lib', '**', '*.rb'))].sort.each { |f| require f }
 Dir[File.expand_path(File.join(File.dirname(__FILE__), 'support', '**', '*.rb'))].sort.each { |f| require f }
